Move get_sell_rate to exchange class
